move direction into its own module
