+++ src/resources/mod.rs
@@ -1,3 +1,4 @@
+pub mod direction;
 pub mod fade;
 pub mod handles;
 pub mod level;
